Add the space replace in the delete script to allow for UTF-8 encoded CSVs
Update the Edit script to include skipping changes for cells that are set to “EMPTY” as this is this is what Mango does for empty strings in the SQL Console.

Update the comments for each script to include the latest SQL queries attached to this ticket.

Update the comments to include an explanation of the logging on/off flag.