Initial upload of event handler
This is a script event handler that should be used with "Point value update" detectors set to "Do not log". This script will log information about the data point being updated, including tags for identification, and calculates the time difference between the point value timestamp sent from the edge and the current timestamp when the cloud server receives the data and processes it. The goal is to help troubleshoot delays in published values by keeping a record of each update.