Initial upload for event-detectors scripts (#11)
* Initial upload for create-event-detectors script

This script only supports the following event detector types:
 *      LOW_LIMIT
 *      HIGH_LIMIT
 *      MULTISTATE_STATE
 *      BINARY_STATE
 *      NO_UPDATE
 *      UPDATE
If additional types are needed, this script can be updated.

* Added feature detection before reloadDataPoint

This latest change tested on Mango 5.0.4 and 5.1.1

* Update create-event-detectors.js

Removed unnecessary comment block and replaced lambda functions to improve readability.

* Delete create_event_detectors/create_event_detectors.js

* Delete create_event_detectors/event_detectors.csv

* Delete create_event_detectors/Event-detector-handlers_to_link/README.md

* Delete create_event_detectors/Event-detector-handlers_to_link/edit_event_detectors.js

* Delete create_event_detectors/Event-detector-handlers_to_link/delete-event-detectors.js

* Delete create_event_detectors/Event-detector-handlers_to_link/create_event_detectors.js

* Uploading new create and edit scripts.

This version of the create script includes a change to set the resetDuration to 0 SECONDS for Mango 5.1
This version of the edit script is new and needs to be reviewed.

* Uploading delete-event-detectors.js

This is a slightly modified version of the existing delete-event-detectors script. The old version was removing spaces from the data rows of the CSV, which was not necessary.

* Delete create-event-detectors.js

Removing the file from this location to add it to a subdirectory for better organization.

* New folder and create-event-detectors.js

* New folder and delete-event-detectors.js

* New folder and edit-event-detectors.js

* Delete create_event_detectors directory

Removing this directory as the name is no longer correct for the contents. A new directory has been added called event_detector_scripts that will replace it.